+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## 1.56.21
+- Added minimumPasswordLength to Organization admin security
+- Added MV84X video resolution details
+- Added NetworkType to OSPF Settings
+- Added MacWhitelistLimit to switch port
+
 ## 1.56.20
 - Added ipsk-with-nac for appliance SSID security type